+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 <!-- markdownlint-disable MD024 -->
 
 [[_TOC_]]
@@ -13,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed -->
 
-## [1.0.0] - #{CREATIONDATE}# - #{MODULEDISPLAYNAME}# module 1st Release
+## [1.0.0] - 2024-11-12 - Azure Container Registry module 1st Release
 
 ### Added
 
-- Added #{MODULEDISPLAYNAME}# module
+- Added Azure Container Registry module